updated README and changed PREC const to usize instead of i32

+++ src/helpers.rs
@@ -33,13 +33,13 @@ pub fn center_in_string(f: String, n: i32) -> String {
 
 #[doc(hidden)]
 pub fn round_and_format(x: f64, latex: bool) -> String {
-    if (x*10f64.powi(PREC-2)).round()/10f64.powi(PREC-2) == 0. && !latex && x != 0. {
+    if (x*10f64.powi(PREC as i32-2)).round()/10f64.powi(PREC as i32-2) == 0. && !latex && x != 0. {
         let mut scientific = format!("{:+e}", x);
         if scientific.chars().nth(0).unwrap() == '+' {
             scientific = scientific[1..].to_string();
         }
         return scientific;
-    } else if (x*10f64.powi(PREC-2)).round()/10f64.powi(PREC-2) == 0. && x != 0. {
+    } else if (x*10f64.powi(PREC as i32-2)).round()/10f64.powi(PREC as i32-2) == 0. && x != 0. {
         let mut scientific = format!("{:+e}", x);
         if scientific.chars().nth(0).unwrap() == '+' {
             scientific = scientific[1..].to_string();
@@ -48,7 +48,7 @@ pub fn round_and_format(x: f64, latex: bool) -> String {
         let right = scientific.split("e").nth(1).unwrap();
         return format!("{}\\cdot 10^{{{}}}", left, right);
     } else {
-        let rounded = (x*10f64.powi(PREC-2)).round()/10f64.powi(PREC-2);
+        let rounded = (x*10f64.powi(PREC as i32-2)).round()/10f64.powi(PREC as i32-2);
         let rounded_string;
         if rounded == 0. && rounded.to_string().len() > 1 {
             rounded_string = rounded.to_string()[1..].to_string();
