+++ src/maths/calculus.rs
@@ -23,7 +23,7 @@ pub fn calculate_integral(expr: &AST, in_terms_of: String, lower_bound: Value, u
                 ub = lb;
                 lb = temp;
             }
-            let dx = (ub-lb)/10f64.powi(PREC-2);
+            let dx = (ub-lb)/10f64.powi(PREC as i32-2);
             let mut sums = vec![];
             let mut b = lb;
             while b < ub {
@@ -67,7 +67,7 @@ pub fn calculate_derivative(expr: &AST, in_terms_of: &str, at: &Value, context: 
             context.add_var(&Variable::new(in_terms_of, vec![at.clone()]));
             let fxs = eval(expr, context)?.to_vec();
             context.remove_var(in_terms_of);
-            context.add_var(&Variable::new(in_terms_of, vec![Value::Scalar(s+10f64.powi(-(PREC)))]));
+            context.add_var(&Variable::new(in_terms_of, vec![Value::Scalar(s+10f64.powi(-(PREC as i32)))]));
             let fxhs = &eval(expr, context)?.to_vec();
             if fxs.len() != fxhs.len() {
                 return Err(EvalError::MathError("Amount of solutions for f(x) and f(x+h) are different!".to_string()));
@@ -81,7 +81,7 @@ pub fn calculate_derivative(expr: &AST, in_terms_of: &str, at: &Value, context: 
                         left: AST::from_value(fxhs[i].clone()),
                         right: AST::from_value(fxs[i].clone())
                     }),
-                    right: AST::from_value(Value::Scalar(10f64.powi(-(PREC))))
+                    right: AST::from_value(Value::Scalar(10f64.powi(-(PREC as i32))))
                 });
                 res.push(eval(&h, &context)?.to_vec());
             }
@@ -108,7 +108,7 @@ pub fn calculate_derivative_newton(expr: &AST, in_terms_of: &str, at: &Value, mu
                 fx = Some(eval(expr, context)?.get(0).unwrap().clone());
                 context.remove_var(in_terms_of);
             }
-            context.add_var(&Variable::new(in_terms_of, vec![Value::Scalar(s+10f64.powi(-(PREC)))]));
+            context.add_var(&Variable::new(in_terms_of, vec![Value::Scalar(s+10f64.powi(-(PREC as i32)))]));
             let fxh = &eval(expr, context)?.get(0).unwrap().clone();
             let h = AST::from_operation(Operation::SimpleOperation {
                 op_type: SimpleOpType::Div,
@@ -117,7 +117,7 @@ pub fn calculate_derivative_newton(expr: &AST, in_terms_of: &str, at: &Value, mu
                     left: AST::from_value(fxh.clone()),
                     right: AST::from_value(fx.clone().unwrap().clone())
                 }),
-                right: AST::from_value(Value::Scalar(10f64.powi(-(PREC))))
+                right: AST::from_value(Value::Scalar(10f64.powi(-(PREC as i32))))
             });
             let res = eval(&h, context)?.get(0).unwrap().clone();
             context.remove_var(in_terms_of);
